+++ test/docs.jl
@@ -1,0 +1,8 @@
+using Enzyme, Test
+using Documenter
+
+DocMeta.setdocmeta!(Enzyme, :DocTestSetup, :(using Enzyme); recursive=true)
+
+@testset "DocTests" begin
+    doctest(Enzyme; manual = false)
+end
