+++ test/ext/logexpfunctions.jl
@@ -1,3 +1,4 @@
+using Enzyme, Test
 using LogExpFunctions
