+++ test/gradjac.jl
@@ -1,10 +1,10 @@
 using Enzyme, Test
-using SparseArrays, StaticArrays
+using LinearAlgebra
+using SparseArrays
 
 using Enzyme: gradient, jacobian
 
-# this is needed but is redundant if included in runtests
-#include("utils.jl")
+@isdefined(UTILS) || include("utils.jl")
 
 
 @testset "Gradient & NamedTuples" begin
@@ -33,7 +33,7 @@ using Enzyme: gradient, jacobian
     @test grad == 14.0
 end
 
-@testset "Gradient & SparseArrays / StaticArrays" begin
+@testset "Gradient & SparseArrays" begin
     x = sparse([5.0, 0.0, 6.0])
     dx = gradient(Reverse, sum, x)
     @test dx isa SparseVector
@@ -43,33 +43,6 @@ end
     dx = gradient(Reverse, sum, x)
     @test dx isa SparseMatrixCSC
     @test dx ≈ [1 0 1]
-
-    x = @SArray [5.0 0.0 6.0]
-    dx = gradient(Reverse, prod, x)
-    @test dx isa SArray
-    @test dx ≈ [0 30 0]
-
-    x = @SVector [1.0, 2.0, 3.0]
-    y = onehot(x)
-    # this should be a very specific type of SArray, but there
-    # is a bizarre issue with older julia versions where it can be MArray
-    @test eltype(y) <: StaticVector
-    @test length(y) == 3
-    @test y[1] == [1.0, 0.0, 0.0]
-    @test y[2] == [0.0, 1.0, 0.0]
-    @test y[3] == [0.0, 0.0, 1.0]
-
-    y = onehot(x, 2, 3)
-    @test eltype(y) <: StaticVector
-    @test length(y) == 2
-    @test y[1] == [0.0, 1.0, 0.0]
-    @test y[2] == [0.0, 0.0, 1.0]
-
-    x = @SArray [5.0 0.0 6.0]
-    dx = gradient(Forward, prod, x)
-    @test dx[1] ≈ 0
-    @test dx[2] ≈ 30
-    @test dx[3] ≈ 0
 end
 
 # these are used in gradient and jacobian tests
