+++ test/runtests.jl
@@ -1,28 +1,15 @@
-# # work around https://github.com/JuliaLang/Pkg.jl/issues/1585
-# using Pkg
-# Pkg.develop(PackageSpec(; path=joinpath(dirname(@__DIR__), "lib", "EnzymeTestUtils")))
-
-using GPUCompiler
-using Enzyme
-using Test
-using FiniteDifferences
-using Aqua
-using SparseArrays
-using StaticArrays
-using Statistics
-using LinearAlgebra
-using InlineStrings
+using Enzyme, Test
 
 using Enzyme_jll
 @info "Testing against" Enzyme_jll.libEnzyme
 
-include("utils.jl")
-
 # Aqua.test_all(Enzyme, unbound_args=false, piracies=false, deps_compat=false, stale_deps=(;:ignore=>[:EnzymeTestUtils]))
 # Aqua.test_all(Enzyme, unbound_args=false, piracies=false, deps_compat=false, stale_deps=(;:ignore=>[:EnzymeTestUtils]))
 
-include("abi.jl")
-include("typetree.jl")
+include("internals/internals.jl")
+include("internals/abi.jl")
+include("internals/usermixed.jl")
+include("internals/typetree.jl")
 
 @static if Enzyme.EnzymeRules.issupported()
     include("rules/rules.jl")
@@ -33,20 +20,22 @@ include("typetree.jl")
     include("rules/internal_rules.jl")
     include("rules/ruleinvalidation.jl")
 end
-@static if !Sys.iswindows()
-    include("blas.jl")
-end
-
-include("internals.jl")
+@static Sys.iswindows() || include("blas.jl")
 
 include("duplicated.jl")
 include("simple.jl")
 include("deferred.jl")
 include("basestd.jl")
 
-include("arrays.jl")
+include("errors.jl")
+include("arrays.jl")  #WARN: this currently broken somehow
 include("memory.jl")
 include("gradjac.jl")
+
+include("mixed.jl")
+include("applyiter.jl")
+include("mixedapplyiter.jl")
+include("misc.jl")
 
 @testset "Threads" begin
     cmd = `$(Base.julia_cmd()) --threads=1 --startup-file=no threads.jl`
@@ -55,19 +44,11 @@ include("gradjac.jl")
    	@test success(pipeline(cmd, stderr=stderr, stdout=stdout))
 end
 
-include("mixed.jl")
-include("applyiter.jl")
-include("mixedapplyiter.jl")
-
 # TEST EXTENSIONS 
-Sys.iswindows() || include("ext/specialfunctions.jl")
+@static Sys.iswindows() || include("ext/specialfunctions.jl")
 include("ext/chainrulescore.jl")
 include("ext/logexpfunctions.jl")
 include("ext/bfloat16s.jl")
+include("ext/staticarrays.jl")
 
-using  Documenter
-DocMeta.setdocmeta!(Enzyme, :DocTestSetup, :(using Enzyme); recursive=true)
-@testset "DocTests" begin
-    doctest(Enzyme; manual = false)
-end
-
+include("docs.jl")
